requirements.txt
langchain
langchain-groq
langgraph
yfinance
pandas
python-dotenv
eventregistry
fastapi
uvicorn[standard]
